Share symbolic solver utilities
