+++ src/compute-engine/symbolic/recurrences.ts
@@ -1,5 +1,14 @@
 import type { Expression } from '../global-types';
 import { isFunction, isSymbol, sym } from '../boxed-expression/type-guards';
+import {
+  appendDistinctRoot,
+  characteristicPolynomial,
+  collectSymbols,
+  freshSymbolName,
+  integrationConstants,
+  rootMultiplicity,
+  solutionRecord,
+} from './solver-utils';
 
 interface RecurrenceProblem {
   equation: Expression;
@@ -17,42 +26,6 @@ function recurrenceProblem(equation: Expression): RecurrenceProblem {
   return recurrence
     ? { equation: recurrence, conditions }
     : { equation, conditions: [] };
-}
-
-function collectSymbols(
-  expr: Expression,
-  symbols = new Set<string>()
-): Set<string> {
-  if (isSymbol(expr)) symbols.add(expr.symbol);
-  if (isFunction(expr)) {
-    for (const op of expr.ops) collectSymbols(op, symbols);
-  }
-  return symbols;
-}
-
-function freshSymbolName(prefix: string, usedSymbols: Set<string>): string {
-  for (let i = 0; ; i++) {
-    const name = i === 0 ? prefix : `${prefix}_${i}`;
-    if (!usedSymbols.has(name)) return name;
-  }
-}
-
-function integrationConstants(
-  equation: Expression,
-  count: number
-): Expression[] {
-  const ce = equation.engine;
-  const usedSymbols = collectSymbols(equation);
-  const result: Expression[] = [];
-
-  for (let i = 1; result.length < count; i++) {
-    const name = `c_${i}`;
-    if (usedSymbols.has(name)) continue;
-    usedSymbols.add(name);
-    result.push(ce.symbol(name));
-  }
-
-  return result;
 }
 
 function functionName(expr: Expression): string | undefined {
@@ -186,64 +159,6 @@ function equationRecurrenceTerms(
   };
 }
 
-function characteristicPolynomial(
-  coefficients: Map<number, Expression>,
-  variable: string,
-  ce: Expression['engine']
-): Expression {
-  const root = ce.symbol(variable);
-  const terms: Expression[] = [];
-
-  for (const [shift, coefficient] of coefficients) {
-    if (coefficient.isSame(0)) continue;
-    if (shift === 0) terms.push(coefficient);
-    else if (shift === 1) terms.push(coefficient.mul(root).simplify());
-    else terms.push(coefficient.mul(root.pow(shift)).simplify());
-  }
-
-  return terms.length === 0 ? ce.Zero : ce.function('Add', terms).simplify();
-}
-
-function isZeroAtRoot(
-  expr: Expression,
-  variable: string,
-  root: Expression
-): boolean {
-  const value = expr.subs({ [variable]: root }).simplify();
-  if (value.isSame(0)) return true;
-  const numeric = value.N();
-  return Math.hypot(numeric.re, numeric.im) < 1e-8;
-}
-
-function rootMultiplicity(
-  polynomial: Expression,
-  variable: string,
-  root: Expression,
-  maxMultiplicity: number
-): number {
-  const ce = polynomial.engine;
-  let multiplicity = 0;
-  let derivative = polynomial;
-
-  while (
-    multiplicity < maxMultiplicity &&
-    isZeroAtRoot(derivative, variable, root)
-  ) {
-    multiplicity += 1;
-    derivative = ce
-      .function('D', [derivative, ce.symbol(variable)])
-      .evaluate()
-      .simplify();
-  }
-
-  return multiplicity;
-}
-
-function appendDistinctRoot(roots: Expression[], root: Expression): void {
-  if (roots.some((other) => root.isSame(other))) return;
-  roots.push(root);
-}
-
 function normalizeShifts(
   terms: RecurrenceTerms
 ): { coefficients: Map<number, Expression>; order: number } | undefined {
@@ -259,20 +174,6 @@ function normalizeShifts(
     coefficients.set(shift - minShift, coefficient.simplify());
 
   return { coefficients, order: maxShift - minShift };
-}
-
-function solutionRecord(
-  result:
-    | null
-    | ReadonlyArray<Expression>
-    | Record<string, Expression>
-    | Array<Record<string, Expression>>
-): Record<string, Expression> | undefined {
-  if (!result) return undefined;
-  if (!Array.isArray(result)) return result as Record<string, Expression>;
-  const [first] = result;
-  if (!first || 'operator' in first) return undefined;
-  return first as Record<string, Expression>;
 }
 
 function conditionEquationForSolution(
